Platform: Move EGLContextExecutor to mirplatform.
Then use this to simplify the Wayland EGL allocator helper; we don't need a separate
`EGLContext` & Wayland executor, we just need somewhere to run cleanup with a current
`EGLContext`. And that's `EGLContextExecutor`.
