+++ src/platform/graphics/egl_context_executor.cpp
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "egl_context_executor.h"
+#include "mir/graphics/egl_context_executor.h"
 #include "mir/renderer/gl/context.h"
 
 namespace mgc = mir::graphics::common;
