+++ src/platforms/x11/graphics/buffer_allocator.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "buffer_allocator.h"
-#include "egl_context_executor.h"
+#include "mir/graphics/egl_context_executor.h"
 #include "shm_buffer.h"
 #include "mir/graphics/egl_extensions.h"
 #include "mir/raii.h"
@@ -225,9 +225,8 @@ std::shared_ptr<mg::Buffer> mgx::BufferAllocator::buffer_from_resource(
         buffer,
         std::move(on_consumed),
         std::move(on_release),
-        ctx,
         *egl_extensions,
-        wayland_executor);
+        egl_delegate);
 }
 
 auto mgx::BufferAllocator::buffer_from_shm(
